fix: move variant resolution client-side via useSearchParams
Receiving searchParams in the page function signature makes the entire
route dynamic under cacheComponents, triggering "uncached data outside
Suspense" errors. Instead, variant resolution now happens client-side:

- New VariantSection client component uses useSearchParams() to read
  variantId and resolve variant selection from the cached product data
- Page no longer touches searchParams at all — stays fully static
- Product data is still server-fetched and cached via "use cache: remote"
- Variant switching is instant (no server round-trip needed)
